[pvr.hts] remove horrible find_if() search and replace with provided data
Since HTSPv6 it turns out we're provided with the ID of any matching DVR
entry. So rather than using a horribly inefficient search algo (especially
when 99.99% of the time there won't be a match) we'll just use the data
provided.

We no longer support anything less than HTSPv10, so no need for a protocol
version check.